requirements.txt
python-telegram-bot==20.3
discord.py==2.6.1
aiohttp
